bet_maker/requirements.txt
aiokafka==0.12.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
click==8.1.7
fastapi==0.115.3
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
kafka-python==2.0.2
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
